requirements.txt
numpy==2.3.4
pandas==2.2.3
scikit-learn==1.7.2
xgboost==2.1.3
shap==0.49.1
optuna==4.1.0
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1
PyYAML==6.0.2
joblib==1.4.2
tqdm==4.66.5
